restructuring for PlatformIO IDE usage instead of Eclipse Arduino or Arduino IDE.
