--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "DCCpp_Uno/LiquidCrystal_PCF8574"]
-	path = DCCpp_Uno/LiquidCrystal_PCF8574
-	url = https://github.com/mathertel/LiquidCrystal_PCF8574
