+++ DCCpp_Uno/.gitignore
@@ -1,2 +1,6 @@
 Release/*
 .ino.cpp
+.pioenvs
+.piolibdeps
+.clang_complete
+.gcc-flags.json
