+++ DCCpp_Uno/platformio.ini
@@ -1,0 +1,21 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; http://docs.platformio.org/page/projectconf.html
+
+
+[env:atmelavr]
+platform=atmelavr
+board=megaatmega2560
+framework=arduino
+lib_deps=
+  EEPROM
+  Wire
+  Ethernet
+  SPI
+  LiquidCrystal_PCF8574
