+++ DCCpp_Uno/src/Accessories.cpp
@@ -31,7 +31,7 @@ variations of the "T" command:
 
   <T>:                         lists all defined turnouts
                                returns: <H ID ADDRESS SUBADDRESS THROW> for each defined turnout or <X> if no turnouts defined
-  
+
 where
 
   ID: the numeric ID (0-32767) of the turnout to control
@@ -62,7 +62,7 @@ the directions of any Turnouts being monitored or controlled by a separate inter
 #include "Accessories.h"
 #include "CommInterface.h"
 #include "SerialCommand.h"
-#include "DCCpp_Uno.h"
+#include "DCCpp.h"
 #include "EEStore.h"
 #include <EEPROM.h>
 
@@ -83,20 +83,20 @@ void Turnout::activate(int s){
 Turnout* Turnout::get(int n){
   Turnout *tt;
   for(tt=firstTurnout;tt!=NULL && tt->data.id!=n;tt=tt->nextTurnout);
-  return(tt); 
+  return(tt);
 }
 ///////////////////////////////////////////////////////////////////////////////
 
 void Turnout::remove(int n){
   Turnout *tt,*pp;
-  
+
   for(tt=firstTurnout;tt!=NULL && tt->data.id!=n;pp=tt,tt=tt->nextTurnout);
 
   if(tt==NULL){
     CommManager::printf("<X>");
     return;
   }
-  
+
   if(tt==firstTurnout)
     firstTurnout=tt->nextTurnout;
   else
@@ -116,7 +116,7 @@ void Turnout::show(int n){
     CommManager::printf("<X>");
     return;
   }
-    
+
   for(tt=firstTurnout;tt!=NULL;tt=tt->nextTurnout){
     if(n==1) {
       CommManager::printf("<H %d %d %d %d>", tt->data.id, tt->data.address, tt->data.subAddress, tt->data.tStatus);
@@ -131,9 +131,9 @@ void Turnout::show(int n){
 void Turnout::parse(const char *c){
   int n,s,m;
   Turnout *t;
-  
+
   switch(sscanf(c,"%d %d %d",&n,&s,&m)){
-    
+
     case 2:                     // argument is string with id number of turnout followed by zero (not thrown) or one (thrown)
       t=get(n);
       if(t!=NULL)
@@ -149,7 +149,7 @@ void Turnout::parse(const char *c){
     case 1:                     // argument is a string with id number only
       remove(n);
     break;
-    
+
     case -1:                    // no arguments
       show(1);                  // verbose show
     break;
@@ -163,22 +163,22 @@ void Turnout::load(){
   Turnout *tt;
 
   for(int i=0;i<EEStore::eeStore->data.nTurnouts;i++){
-    EEPROM.get(EEStore::pointer(),data);  
+    EEPROM.get(EEStore::pointer(),data);
     tt=create(data.id,data.address,data.subAddress);
     tt->data.tStatus=data.tStatus;
     tt->num=EEStore::pointer();
     EEStore::advance(sizeof(tt->data));
-  }  
+  }
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 
 void Turnout::store(){
   Turnout *tt;
-  
+
   tt=firstTurnout;
   EEStore::eeStore->data.nTurnouts=0;
-  
+
   while(tt!=NULL){
     tt->num=EEStore::pointer();
     EEPROM.put(EEStore::pointer(),tt->data);
@@ -186,13 +186,13 @@ void Turnout::store(){
     tt=tt->nextTurnout;
     EEStore::eeStore->data.nTurnouts++;
   }
-  
+
 }
 ///////////////////////////////////////////////////////////////////////////////
 
 Turnout *Turnout::create(int id, int add, int subAdd, int v){
   Turnout *tt;
-  
+
   if(firstTurnout==NULL){
     firstTurnout=(Turnout *)calloc(1,sizeof(Turnout));
     tt=firstTurnout;
@@ -209,7 +209,7 @@ Turnout *Turnout::create(int id, int add, int subAdd, int v){
       CommManager::printf("<X>");
     return(tt);
   }
-  
+
   tt->data.id=id;
   tt->data.address=add;
   tt->data.subAddress=subAdd;
@@ -217,11 +217,9 @@ Turnout *Turnout::create(int id, int add, int subAdd, int v){
   if(v==1)
     CommManager::printf("<O>");
   return(tt);
-  
+
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 
 Turnout *Turnout::firstTurnout=NULL;
-
-
