+++ DCCpp_Uno/src/CurrentMonitor.cpp
@@ -7,7 +7,7 @@ Part of DCC++ BASE STATION for the Arduino
 
 **********************************************************************/
 
-#include "DCCpp_Uno.h"
+#include "DCCpp.h"
 #include "CurrentMonitor.h"
 #include "CommInterface.h"
