+++ DCCpp_Uno/src/EEStore.cpp
@@ -7,7 +7,7 @@ Part of DCC++ BASE STATION for the Arduino
 
 **********************************************************************/
 
-#include "DCCpp_Uno.h"
+#include "DCCpp.h"
 #include "EEStore.h"
 #include "Accessories.h"
 #include "Sensor.h"
@@ -18,36 +18,36 @@ Part of DCC++ BASE STATION for the Arduino
 
 void EEStore::init(){
 
-  
+
   eeStore=(EEStore *)calloc(1,sizeof(EEStore));
 
-  EEPROM.get(0,eeStore->data);                                       // get eeStore data 
-  
+  EEPROM.get(0,eeStore->data);                                       // get eeStore data
+
   if(strncmp(eeStore->data.id,EESTORE_ID,sizeof(EESTORE_ID))!=0){    // check to see that eeStore contains valid DCC++ ID
     sprintf(eeStore->data.id,EESTORE_ID);                           // if not, create blank eeStore structure (no turnouts, no sensors) and save it back to EEPROM
     eeStore->data.nTurnouts=0;
     eeStore->data.nSensors=0;
     eeStore->data.nOutputs=0;
-    EEPROM.put(0,eeStore->data);    
+    EEPROM.put(0,eeStore->data);
   }
-  
+
   reset();            // set memory pointer to first free EEPROM space
   Turnout::load();    // load turnout definitions
   Sensor::load();     // load sensor definitions
   Output::load();     // load output definitions
-  
+
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 
 void EEStore::clear(){
-    
+
   sprintf(eeStore->data.id,EESTORE_ID);                           // create blank eeStore structure (no turnouts, no sensors) and save it back to EEPROM
   eeStore->data.nTurnouts=0;
   eeStore->data.nSensors=0;
   eeStore->data.nOutputs=0;
-  EEPROM.put(0,eeStore->data);    
-  
+  EEPROM.put(0,eeStore->data);
+
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -55,9 +55,9 @@ void EEStore::clear(){
 void EEStore::store(){
   reset();
   Turnout::store();
-  Sensor::store();  
-  Output::store();  
-  EEPROM.put(0,eeStore->data);    
+  Sensor::store();
+  Output::store();
+  EEPROM.put(0,eeStore->data);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -80,4 +80,3 @@ int EEStore::pointer(){
 
 EEStore *EEStore::eeStore=NULL;
 int EEStore::eeAddress=0;
-
